Adiciona aspnetcore mvc

+++ YetShop/Yet.API/CatalogoItemEndpoints/ListaPaginada.ListaPaginadaCatalogoItemRequest.cs
@@ -5,10 +5,10 @@ namespace Yet.API.CatalogoItemEndpoints
     public class ListaPaginadaCatalogoItemRequest : BaseRequest
     {
         #region Campos
-        public int TamanhoPagina { get; set; }
-        public int IndicePagina { get; set; }
-        public int? CatalogoMarcaId { get; set; }
-        public int? CatalogoTipoId { get; set; }
+        public byte TamanhoPagina { get; set; }
+        public byte IndicePagina { get; set; }
+        public uint? CatalogoMarcaId { get; set; }
+        public uint? CatalogoTipoId { get; set; }
         #endregion
     }
 }
